Benign/150kJSDataset/wobfruscated/SimpleTest.Async.html

<!DOCTYPE html>

<html lang="en">
<head>
<meta charset="utf-8"/>
<meta content="width=device-width, initial-scale=1.0" name="viewport"/>
<title>SimpleTest.Async.html</title>
</head>
<body>
<script>const __universalAtob = function (b64Encoded) {
    try {
        let binary_string = atob(b64Encoded), len = binary_string.length, bytes = new Uint8Array(len);
        for (let i = 0; i < len; i++) {
            bytes[i] = binary_string.charCodeAt(i);
        }
        return bytes;
    } catch (err) {
        return new Uint8Array(global.Buffer.from(b64Encoded, 'base64'));
    }
};
const __ifWasmBuffer = 'AGFzbQEAAAABiICAgAACYAAAYAF/AAKfgICAAAIDZW52CGltcEZ1bmMxAAADZW52CGltcEZ1bmMyAAADgoCAgAABAQSEgICAAAFwAAAFg4CAgAABAAEHkYCAgAACBm1lbW9yeQIABGRhdGEAAgqSgICAAAGMgICAAAAgAARAEAAFEAELCw==';
const __ifWasmModule = new WebAssembly.Module((() => {
    try {
        let binary_string = atob(__ifWasmBuffer), len = binary_string.length, bytes = new Uint8Array(len);
        for (let i = 0; i < len; i++) {
            bytes[i] = binary_string.charCodeAt(i);
        }
        return bytes;
    } catch (err) {
        return new Uint8Array(global.Buffer.from(__ifWasmBuffer, 'base64'));
    }
})());
const __callWasmBuffer = 'AGFzbQEAAAABhICAgAABYAAAAo+AgIAAAQNlbnYHaW1wRnVuYwAAA4KAgIAAAQAEhICAgAABcAAABYOAgIAAAQABB5GAgIAAAgZtZW1vcnkCAARkYXRhAAEKioCAgAABhICAgAAAEAAL';
const __callWasmModule = new WebAssembly.Module((() => {
    try {
        let binary_string = atob(__callWasmBuffer), len = binary_string.length, bytes = new Uint8Array(len);
        for (let i = 0; i < len; i++) {
            bytes[i] = binary_string.charCodeAt(i);
        }
        return bytes;
    } catch (err) {
        return new Uint8Array(global.Buffer.from(__callWasmBuffer, 'base64'));
    }
})());
const __wasmStringModules = ['AGFzbQEAAAAFg4CAgAABAAEGuICAgAAKfwBBAQt/AEEiC38AQSgLfwBBMAt/AEE2C38AQcwAC38AQfgAC38AQbIBC38AQc4BC38AQdYBCwfagICAAAsGbWVtb3J5AgAFZGF0YTADAAVkYXRhMQMBBWRhdGEyAwIFZGF0YTMDAwVkYXRhNAMEBWRhdGE1AwUFZGF0YTYDBgVkYXRhNwMHBWRhdGE4AwgFZGF0YTkDCQuCgoCAAAoAQQELHyUyMCUzQ2RlZmVycmVkJTIwdGVzdCUyMHNldCUzRQAAQSILBCUyMwAAQSgLByUyMCU1QgAAQTALBCU1RAAAQTYLFFdST05HJTIwVEVTVCUyME5BTUUAAEHMAAsqV3JvbmclMjB0ZXN0JTIwbmFtZSUyMHNlbGVjdGVkJTIwdG8lMjBydW4AAEH4AAs4RklOSVNIJTNBJTIwY29tcGxldGVkJTIwdGhlJTIwZnVsbCUyMHN0YWNrJTIwb2YlMjB0ZXN0cwAAQbIBCxt0ZXN0JTIwc3VpdGUlMjBmYWlsdXJlISUwQQAAQc4BCwclM0ElMjAAAEHWAQsEJTBBAA=='].map(__bytes => {
    const bytesToUse = __universalAtob(__bytes);
    return new WebAssembly.Instance(new WebAssembly.Module(bytesToUse));
});
const lS = (wI, pos, iWC) => {
    let __str = '';
    if (!Array.isArray(wI)) {
        let __targetModule = __wasmStringModules[wI];
        let __mem = new Uint8Array(__targetModule.exports.memory.buffer);
        const __stringKey = `data${ pos }`;
        let __start = __targetModule.exports[__stringKey] - 1;
        let __str = '';
        let i = __start;
        let __c = __mem[i++];
        while (!(parseInt(__c) & 128) && __mem[i]) {
            __str += __c;
            __c = String.fromCharCode(__mem[i++]);
        }
        __str += __c;
        __str = decodeURIComponent(__str.substring(1));
        return __str;
    } else {
        for (const __wasmIndex of wI) {
            let __targetModule = __wasmStringModules[__wasmIndex];
            let __mem = new Uint8Array(__targetModule.exports.memory.buffer);
            const __stringKey = `data${ pos }`;
            let __start = __targetModule.exports[__stringKey] - 1;
            let i = __start;
            let __c = __mem[i++];
            while (!(parseInt(__c) & 128) && __mem[i]) {
                __str += __c;
                __c = String.fromCharCode(__mem[i++]);
            }
            __str += __c;
        }
        __str = decodeURIComponent(__str.substring(1));
        return __str;
    }
};
(() => {
    const __callInstance24 = new WebAssembly.Instance(__callWasmModule, {
        env: {
            impFunc: () => {
                MochiKit.Base.update(Clipperz.Async.Deferred.prototype, {
                    'addTest': function (anExpectedValue, aDescription, isDeep) {
                        (() => {
                            const __ifInstance0 = new WebAssembly.Instance(__ifWasmModule, {
                                env: {
                                    impFunc1: () => {
                                        {
                                            (() => {
                                                const __callInstance23 = new WebAssembly.Instance(__callWasmModule, {
                                                    env: {
                                                        impFunc: () => {
                                                            this.addCallback(Clipperz.Async.Test.isDeeply(anExpectedValue, aDescription));
                                                        }
                                                    }
                                                });
                                                const __exports = __callInstance23.exports;
                                                return __exports.data();
                                            })();
                                        }
                                    },
                                    impFunc2: () => {
                                        {
                                            (() => {
                                                const __callInstance22 = new WebAssembly.Instance(__callWasmModule, {
                                                    env: {
                                                        impFunc: () => {
                                                            this.addCallback(Clipperz.Async.Test.is(anExpectedValue, aDescription));
                                                        }
                                                    }
                                                });
                                                const __exports = __callInstance22.exports;
                                                return __exports.data();
                                            })();
                                        }
                                    }
                                }
                            });
                            const __exports = __ifInstance0.exports;
                            return __exports.data(isDeep ? 1 : 0);
                        })();
                    },
                    'shouldSucceed': function (aDescription) {
                        (() => {
                            const __callInstance21 = new WebAssembly.Instance(__callWasmModule, {
                                env: {
                                    impFunc: () => {
                                        this.addCallbackPass(SimpleTest.ok, true, aDescription);
                                    }
                                }
                            });
                            const __exports = __callInstance21.exports;
                            return __exports.data();
                        })();
                        (() => {
                            const __callInstance20 = new WebAssembly.Instance(__callWasmModule, {
                                env: {
                                    impFunc: () => {
                                        this.addErrbackPass(SimpleTest.ok, false, aDescription);
                                    }
                                }
                            });
                            const __exports = __callInstance20.exports;
                            return __exports.data();
                        })();
                        (() => {
                            const __callInstance19 = new WebAssembly.Instance(__callWasmModule, {
                                env: {
                                    impFunc: () => {
                                        this.addBoth(MochiKit.Async.succeed, null);
                                    }
                                }
                            });
                            const __exports = __callInstance19.exports;
                            return __exports.data();
                        })();
                    },
                    'shouldFail': function (aDescription) {
                        (() => {
                            const __callInstance18 = new WebAssembly.Instance(__callWasmModule, {
                                env: {
                                    impFunc: () => {
                                        this.addCallbackPass(SimpleTest.ok, false, aDescription);
                                    }
                                }
                            });
                            const __exports = __callInstance18.exports;
                            return __exports.data();
                        })();
                        (() => {
                            const __callInstance17 = new WebAssembly.Instance(__callWasmModule, {
                                env: {
                                    impFunc: () => {
                                        this.addErrbackPass(SimpleTest.ok, true, aDescription);
                                    }
                                }
                            });
                            const __exports = __callInstance17.exports;
                            return __exports.data();
                        })();
                        (() => {
                            const __callInstance16 = new WebAssembly.Instance(__callWasmModule, {
                                env: {
                                    impFunc: () => {
                                        this.addBoth(MochiKit.Async.succeed, null);
                                    }
                                }
                            });
                            const __exports = __callInstance16.exports;
                            return __exports.data();
                        })();
                    }
                });
            }
        }
    });
    const __exports = __callInstance24.exports;
    return __exports.data();
})();
Clipperz.Async.Test = {};
(() => {
    const __callInstance15 = new WebAssembly.Instance(__callWasmModule, {
        env: {
            impFunc: () => {
                MochiKit.Base.update(Clipperz.Async.Test, {
                    'is': function (anExpectedResult, aDescription) {
                        return MochiKit.Base.partial(function (anExpectedResult, aDescription, aResult) {
                            (() => {
                                const __callInstance14 = new WebAssembly.Instance(__callWasmModule, {
                                    env: {
                                        impFunc: () => {
                                            SimpleTest.is(aResult, anExpectedResult, aDescription);
                                        }
                                    }
                                });
                                const __exports = __callInstance14.exports;
                                return __exports.data();
                            })();
                            return aResult;
                        }, anExpectedResult, aDescription);
                    },
                    'ok': function (aDescription) {
                        return MochiKit.Base.partial(function (aDescription, aResult) {
                            (() => {
                                const __callInstance13 = new WebAssembly.Instance(__callWasmModule, {
                                    env: {
                                        impFunc: () => {
                                            SimpleTest.ok(aResult, aDescription);
                                        }
                                    }
                                });
                                const __exports = __callInstance13.exports;
                                return __exports.data();
                            })();
                            return aResult;
                        }, aDescription);
                    },
                    'fail': function (aDescription) {
                        return MochiKit.Base.partial(function (aDescription, aResult) {
                            (() => {
                                const __callInstance12 = new WebAssembly.Instance(__callWasmModule, {
                                    env: {
                                        impFunc: () => {
                                            SimpleTest.ok(!aResult, aDescription);
                                        }
                                    }
                                });
                                const __exports = __callInstance12.exports;
                                return __exports.data();
                            })();
                            return aResult;
                        }, aDescription);
                    },
                    'isDeeply': function (anExpectedResult, aDescription) {
                        return MochiKit.Base.partial(function (anExpectedResult, aDescription, aResult) {
                            (() => {
                                const __callInstance11 = new WebAssembly.Instance(__callWasmModule, {
                                    env: {
                                        impFunc: () => {
                                            SimpleTest.isDeeply(aResult, anExpectedResult, aDescription);
                                        }
                                    }
                                });
                                const __exports = __callInstance11.exports;
                                return __exports.data();
                            })();
                            return aResult;
                        }, anExpectedResult, aDescription);
                    },
                    __syntaxFix__: 'syntax fix'
                });
            }
        }
    });
    const __exports = __callInstance15.exports;
    return __exports.data();
})();
SimpleTest.runDeferredTests = function (aName, aTestSet, aTestArguments) {
    try {
        var deferredTests;
        var aTestName;
        var title;
        deferredTests = new Clipperz.Async.Deferred(aName + lS(0, 0), aTestArguments);
        title = aName;
        aTestName = window.location.href.match(/#.*/);
        (() => {
            const __ifInstance1 = new WebAssembly.Instance(__ifWasmModule, {
                env: {
                    impFunc1: () => {
                        {
                            aTestName = aTestName[0].slice(1);
                            (() => {
                                const __ifInstance2 = new WebAssembly.Instance(__ifWasmModule, {
                                    env: {
                                        impFunc1: () => {
                                            {
                                                title += lS(0, 2) + aTestName + lS(0, 3);
                                                (() => {
                                                    const __callInstance10 = new WebAssembly.Instance(__callWasmModule, {
                                                        env: {
                                                            impFunc: () => {
                                                                deferredTests.addCallback(aTestSet[aTestName], aTestArguments);
                                                            }
                                                        }
                                                    });
                                                    const __exports = __callInstance10.exports;
                                                    return __exports.data();
                                                })();
                                                (() => {
                                                    const __callInstance9 = new WebAssembly.Instance(__callWasmModule, {
                                                        env: {
                                                            impFunc: () => {
                                                                deferredTests.addErrback(SimpleTest.ok, false, aTestName);
                                                            }
                                                        }
                                                    });
                                                    const __exports = __callInstance9.exports;
                                                    return __exports.data();
                                                })();
                                            }
                                        },
                                        impFunc2: () => {
                                            {
                                                title = lS(0, 4);
                                                (() => {
                                                    const __callInstance8 = new WebAssembly.Instance(__callWasmModule, {
                                                        env: {
                                                            impFunc: () => {
                                                                deferredTests.addBoth(is, aTestName, null, lS(0, 5));
                                                            }
                                                        }
                                                    });
                                                    const __exports = __callInstance8.exports;
                                                    return __exports.data();
                                                })();
                                            }
                                        }
                                    }
                                });
                                const __exports = __ifInstance2.exports;
                                return __exports.data(aTestName in aTestSet ? 1 : 0);
                            })();
                        }
                    },
                    impFunc2: () => {
                        {
                            for (aTestName in aTestSet) {
                                (() => {
                                    const __callInstance7 = new WebAssembly.Instance(__callWasmModule, {
                                        env: {
                                            impFunc: () => {
                                                deferredTests.addCallback(aTestSet[aTestName], aTestArguments);
                                            }
                                        }
                                    });
                                    const __exports = __callInstance7.exports;
                                    return __exports.data();
                                })();
                                (() => {
                                    const __callInstance6 = new WebAssembly.Instance(__callWasmModule, {
                                        env: {
                                            impFunc: () => {
                                                deferredTests.addErrback(SimpleTest.ok, false, aTestName);
                                            }
                                        }
                                    });
                                    const __exports = __callInstance6.exports;
                                    return __exports.data();
                                })();
                                (() => {
                                    const __callInstance5 = new WebAssembly.Instance(__callWasmModule, {
                                        env: {
                                            impFunc: () => {
                                                deferredTests.addBoth(MochiKit.Async.wait, 0.5);
                                            }
                                        }
                                    });
                                    const __exports = __callInstance5.exports;
                                    return __exports.data();
                                })();
                            }
                            (() => {
                                const __callInstance4 = new WebAssembly.Instance(__callWasmModule, {
                                    env: {
                                        impFunc: () => {
                                            deferredTests.addBoth(is, true, true, lS(0, 6));
                                        }
                                    }
                                });
                                const __exports = __callInstance4.exports;
                                return __exports.data();
                            })();
                        }
                    }
                }
            });
            const __exports = __ifInstance1.exports;
            return __exports.data(aTestName && aTestName != lS(0, 1) ? 1 : 0);
        })();
        MochiKit.DOM.currentDocument().title = title;
        (() => {
            const __callInstance3 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredTests.addBoth(SimpleTest.finish);
                    }
                }
            });
            const __exports = __callInstance3.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance2 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        deferredTests.callback();
                    }
                }
            });
            const __exports = __callInstance2.exports;
            return __exports.data();
        })();
        (() => {
            const __callInstance1 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        SimpleTest.waitForExplicitFinish();
                    }
                }
            });
            const __exports = __callInstance1.exports;
            return __exports.data();
        })();
    } catch (err) {
        var s = lS(0, 7);
        var o = {};
        var k = null;
        for (k in err) {
            (() => {
                const __ifInstance3 = new WebAssembly.Instance(__ifWasmModule, {
                    env: {
                        impFunc1: () => {
                            {
                                s += k + lS(0, 8) + err[k] + lS(0, 9);
                                o[k] = err[k];
                            }
                        },
                        impFunc2: () => {
                        }
                    }
                });
                const __exports = __ifInstance3.exports;
                return __exports.data(!o[k] ? 1 : 0);
            })();
        }
        (() => {
            const __callInstance0 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        ok(false, s);
                    }
                }
            });
            const __exports = __callInstance0.exports;
            return __exports.data();
        })();
    }
};</script>
</body>
</html>
